EZ195/TheLink#3
[이지원]

- 프로필 수정 시 이미지를 업로드하지 않으면 기존의 이미지 파일이 삭제되는 오류 해결

+++ src/main/java/com/ezone/link/user/profile/bo/ProfileBO.java
@@ -36,10 +36,16 @@ public class ProfileBO {
 	public int updateProfile(int userId, String nickname, String introduce, MultipartFile file) {
 		
 		Profile profile = this.getUserProfile(userId);
+		String filepath = "";
 		
-		FileManagerService.removeFile(profile.getProfileImagePath());
+		if (file == null) {
+			filepath = profile.getProfileImagePath();
+		}
 		
-		String filepath = FileManagerService.saveFile(userId, file);
+		else {
+			FileManagerService.removeFile(profile.getProfileImagePath());
+			filepath = FileManagerService.saveFile(userId, file);
+		}
 		
 		return profileDAO.updateUserProfile(userId, nickname, introduce, filepath);
 	}
